+++ elena/domain/services/generic_bot.py
@@ -70,28 +70,10 @@ class GenericBot(Bot):
         self.exchange_manager = exchange_manager
         self._update_orders_status()
 
-    #  ---- StatusManager
-
-    def _new_trade_by_order(self, order: Order):
-        # All Trades start/"born" here...
-        new_trade = Trade(
-            exchange_id=self.exchange.id,
-            bot_id=self.id,
-            strategy_id=self.bot_config.strategy_id,
-            pair=self.pair,
-            size=order.amount,
-            entry_order_id=order.id,
-            entry_price=order.average,
-            entry_time=order.timestamp,
-            exit_order_id='0',  # TODO define an OrderId.Null
-            exit_price=0,
-        )
-        self.status.active_trades.append(new_trade)
-
     def new_trade_manual(self, size: float, entry_price: float, exit_order_id, exit_price: float):
         new_trade = Trade(
             exchange_id=self.exchange.id,
-            bot_id=self.id,
+            bot_id=self.bot_id,
             strategy_id=self.bot_config.strategy_id,
             pair=self.pair,
             size=size,
@@ -100,144 +82,7 @@ class GenericBot(Bot):
             exit_order_id=exit_order_id,
             exit_price=exit_price,
         )
-        self.status.active_trades.append(new_trade)
-
-    def _close_individual_trade_on_new_order(self, trade: Trade, order: Order, amount_to_close: float) -> float:
-        if trade.size <= amount_to_close:
-            self.status.active_trades.remove(trade)
-            trade.exit_time = order.timestamp
-            trade.exit_price = order.average
-            self.status.closed_trades.append(trade)
-            return amount_to_close - trade.size
-        else:
-            self._logger.error(f"Amount to close is insufficient for trade id:{trade.id} size: {trade.size} on order {order.id} size: {order.amount} with pending to close {amount_to_close}")
-            return amount_to_close
-
-    def _close_trades_on_new_updated_order(self, order: Order):
-        amount_to_close = order.amount
-        # check trades with an exit order id
-        for trade in self.status.active_trades:
-            if trade.exit_order_id == order.id and amount_to_close > 0:
-                amount_to_close = self._close_individual_trade_on_new_order(trade, order, amount_to_close)
-
-        # check trades without an exit order id
-        if amount_to_close > 0:
-            for trade in self.status.active_trades:
-                if trade.exit_order_id == '0' and amount_to_close > 0: # TODO define an OrderId.Null
-                    amount_to_close = self._close_individual_trade_on_new_order(trade, order, amount_to_close)
-
-        # close trade even with a different order_id
-        if amount_to_close > 0:
-            self._logger.warning("The order size is bigger than the trades with an explicit order_id or a blank order_id")
-            for trade in self.status.active_trades:
-                if amount_to_close > 0:
-                    amount_to_close = self._close_individual_trade_on_new_order(trade, order, amount_to_close)
-
-        if amount_to_close > 0:  # TODO: how to pass min_amount?
-            self._logger.error("The order size is bigger than any trade. {amount_to_close} left to close of {order.amount}.")
-
-    def _register_new_order_on_trades(self, order: Order):
-        if order is None:
-            raise "Order can't be None"
-
-        # TODO OrderStatusType.canceled or rejected
-        if order.status == OrderStatusType.canceled or order.status == OrderStatusType.rejected:
-            raise "Order condition unhandled (canceled or rejected)"
-
-        if order.side == OrderSide.buy:
-            self._new_trade_by_order(order)
-            # TODO: budget.lock
-            if order.status == OrderStatusType.closed:
-                self.status.archived_orders.append(order)
-            else:  # open & partials are active, budget is lock equally.
-                self.status.active_orders.append(order)
-
-        elif order.side == OrderSide.sell:
-            if order.status == OrderStatusType.closed:
-                # TODO: budget.unlock
-                self._close_trades_on_new_updated_order(order)
-                self.status.archived_orders.append(order)
-            else:
-                # stop loss => if order.stop_price and order.stop_price > 0:
-                # TODO: budget.unlock (partial) ???
-                self.status.active_orders.append(order)
-        else:
-            raise "Order condition unhandled (OrderSide)"
-
-    def _update_trades_on_update_orders(self, order: Order):
-        if order is None:
-            raise "Order can't be None"
-
-        if order.status == OrderStatusType.rejected:
-            # on rejected
-            #   => problem... the market could be closed.
-            #   keep processing trades
-            self._logger.error(f"Order {order.id} was rejected. Bot: {self.id}")
-            self._notifications_manager.high(f"Order {order.id} was rejected. Bot: {self.id}")
-
-        if order.side == OrderSide.buy:
-            if order.status == OrderStatusType.closed:
-                # on buy, close
-                #   update trade.order status (done in _update_orders_status)
-                # nothing to do here
-                pass
-            elif order.status == OrderStatusType.canceled or order.status == OrderStatusType.rejected:
-                # on buy, cancel or rejected
-                #   archive trade
-                trade_found = False
-                for trade in self.status.active_trades:
-                    if trade.entry_order_id == order.id:
-                        self.status.active_trades.remove(trade)
-                        trade.exit_time = order.timestamp
-                        trade.exit_price = 0
-                        trade.exit_order_id = order.status
-                        self.status.closed_trades.append(trade)
-                        trade_found = True
-                if not trade_found:
-                    self._logger.error(f"Order {order.id} canceled or rejected not found in trades. Bot: {self.id}")
-
-                # TODO: budget.unlock
-            elif order.status == OrderStatusType.open:  # open & partials are active, budget is lock equally.
-                # on buy, open or partial
-                #   update trade.order status (done in _update_orders_status)
-                # nothing to do here
-                pass
-            else:
-                raise "Order condition unhandled (OrderSide.buy)"
-
-        elif order.side == OrderSide.sell:
-            if order.status == OrderStatusType.closed:
-                # on sell, close
-                #   close trade... _close_trades_on_new_updated_order?
-                self._close_trades_on_new_updated_order(order)
-                # TODO: budget.unlock
-            elif order.status == OrderStatusType.canceled or order.status == OrderStatusType.rejected:
-                # on sell, cancel or rejected
-                #   do nothing unless partial
-                # TODO: if a partial order is cancelled... what should be done?
-                pass
-            elif order.status == OrderStatusType.open:
-                # on sell, partial
-                #   update trade.order status (done in _update_orders_status)
-                #   budget.unlock(partial)? => risky
-                # stop loss => if order.stop_price and order.stop_price > 0:
-                # TODO: budget.unlock (partial) ???
-                pass
-            else:
-                raise "Order condition unhandled (OrderSide.sell)"
-        else:
-            raise "Order condition unhandled (OrderSide)"
-
-    def _archive_order_on_cancel(self, order: Order):
-        found_order = False
-        for loop_order in self.status.active_orders:
-            if loop_order.id == order.id:
-                self.status.active_orders.remove(loop_order)
-                found_order = True
-        if not found_order:
-            self._logger.error(f"Order {order.id} canceled but not found in active_orders. Bot: {self.id}")
-        self.status.archived_orders.append(order)
-        self._update_trades_on_update_orders(order)  # TODO: check
+        self.active_trades.append(new_trade)
 
     def _update_orders_status(self) -> BotStatus:
         # orders
@@ -249,7 +94,7 @@ class GenericBot(Bot):
             updated_order = self.fetch_order(order.id)
 
             if updated_order:
-                self._update_trades_on_update_orders(updated_order)
+                self.status.update_trades_on_update_orders(updated_order)
                 if updated_order.status in [OrderStatusType.closed, OrderStatusType.canceled, OrderStatusType.rejected]:
                     # move to archived
                     self.status.archived_orders.append(updated_order)
@@ -352,7 +197,7 @@ class GenericBot(Bot):
             )
             self._metrics_manager.counter(Metric.ORDER_CANCELLED, self.bot_config)
             self._notifications_manager.medium(f"Order {order_id} was cancelled")
-            self._archive_order_on_cancel(order)
+            self.status.archive_order_on_cancel(order)
             return order
         except Exception as err:
             print(f"Error cancelling order: {err}")
@@ -390,7 +235,7 @@ class GenericBot(Bot):
             )
             self._logger.info("Placed market stop loss: %s", order)
 
-            self._register_new_order_on_trades(order)
+            self.status.register_new_order_on_trades(order)
             return order
         except Exception as err:
             print(f"Error creating stop loss: {err}")
@@ -419,7 +264,7 @@ class GenericBot(Bot):
             )
             self._logger.info("Placed market buy: %s", order)
 
-            self._register_new_order_on_trades(order)
+            self.status.register_new_order_on_trades(order)
             return order
         except Exception as err:
             print(f"Error creating market buy order: {err}")
@@ -441,7 +286,7 @@ class GenericBot(Bot):
             )
             self._logger.info("Placed market sell: %s", order)
 
-            self._register_new_order_on_trades(order)
+            self.status.register_new_order_on_trades(order)
             return order
         except Exception as err:
             print(f"Error creating market sell order: {err}")
